Added LayerSprite, first approximation

+++ model/stage/Stage.cpp
@@ -1,7 +1,7 @@
 #include "Stage.h"
 
 
-Stage::Stage(std::list<Layer*>* passed_layers, float passed_width, float passed_heigh, float passed_floor) {
+Stage::Stage(std::list<Layer*>* passed_layers, int passed_width, int passed_heigh, int passed_floor) {
     this->layers = passed_layers;
     this->width = passed_width;
     this->height = passed_heigh;
